refactor: colocate per-concern tests + silence two re-export warnings (#190)
Moves the instability/blast_radius integration tests from tests.rs into
metrics.rs (where they exercise compute_hotspots directly) and the
independence_* tests into independence.rs. tests.rs now focuses on
coupling/score/circular/audit_with_settings integration only, dropping
from 666 to 451 LOC.

Also adds #[allow(unused_imports)] to the analyzer::TopAction and
analyzer::CrossModuleViolation re-exports — they're part of the
analyzer's public surface (they were pub before the split) but no
in-tree caller currently uses them, which Rust's unused-import lint
flags after re-export.
