+++ src/analyzer/metrics.rs
@@ -87,3 +87,143 @@ pub fn compute_hotspots(modules: &[Module], dependencies: &[Dependency]) -> Vec<
     hotspots.sort_by_key(|h| std::cmp::Reverse(h.fan_in));
     hotspots
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::core::ModuleType;
+
+    fn make_module(id: &str, path: &str) -> Module {
+        Module {
+            id: id.to_string(),
+            snapshot_id: "snap".to_string(),
+            parent_id: None,
+            name: std::path::Path::new(path)
+                .file_name()
+                .unwrap()
+                .to_string_lossy()
+                .to_string(),
+            path: path.to_string(),
+            module_type: ModuleType::File,
+            depth: std::path::Path::new(path).components().count() as i32,
+        }
+    }
+
+    #[test]
+    fn instability_pure_consumer() {
+        let modules = vec![
+            make_module("a1", "src/app/main.rs"),
+            make_module("l1", "src/lib/core.rs"),
+        ];
+        let deps = vec![Dependency {
+            from_module_id: "a1".to_string(),
+            to_module_id: "l1".to_string(),
+            line_number: 1,
+        }];
+        let hotspots = compute_hotspots(&modules, &deps);
+        let a1 = hotspots
+            .iter()
+            .find(|h| h.path == "src/app/main.rs")
+            .unwrap();
+        assert_eq!(a1.fan_out, 1);
+        assert_eq!(a1.fan_in, 0);
+        assert!((a1.instability - 1.0).abs() < 0.01);
+    }
+
+    #[test]
+    fn instability_pure_provider() {
+        let modules = vec![
+            make_module("a1", "src/app/main.rs"),
+            make_module("l1", "src/lib/core.rs"),
+        ];
+        let deps = vec![Dependency {
+            from_module_id: "a1".to_string(),
+            to_module_id: "l1".to_string(),
+            line_number: 1,
+        }];
+        let hotspots = compute_hotspots(&modules, &deps);
+        let l1 = hotspots
+            .iter()
+            .find(|h| h.path == "src/lib/core.rs")
+            .unwrap();
+        assert_eq!(l1.fan_in, 1);
+        assert_eq!(l1.fan_out, 0);
+        assert!((l1.instability - 0.0).abs() < 0.01);
+    }
+
+    #[test]
+    fn instability_balanced() {
+        let modules = vec![
+            make_module("a", "src/a.rs"),
+            make_module("b", "src/b.rs"),
+            make_module("c", "src/c.rs"),
+        ];
+        let deps = vec![
+            Dependency {
+                from_module_id: "a".to_string(),
+                to_module_id: "b".to_string(),
+                line_number: 1,
+            },
+            Dependency {
+                from_module_id: "b".to_string(),
+                to_module_id: "c".to_string(),
+                line_number: 1,
+            },
+        ];
+        let hotspots = compute_hotspots(&modules, &deps);
+        let b = hotspots.iter().find(|h| h.path == "src/b.rs").unwrap();
+        assert_eq!(b.fan_in, 1);
+        assert_eq!(b.fan_out, 1);
+        assert!((b.instability - 0.5).abs() < 0.01);
+    }
+
+    #[test]
+    fn blast_radius_leaf_is_zero() {
+        let modules = vec![make_module("a", "src/a.rs"), make_module("b", "src/b.rs")];
+        let deps = vec![Dependency {
+            from_module_id: "a".to_string(),
+            to_module_id: "b".to_string(),
+            line_number: 1,
+        }];
+        let hotspots = compute_hotspots(&modules, &deps);
+        let a = hotspots.iter().find(|h| h.path == "src/a.rs").unwrap();
+        assert_eq!(a.blast_radius, 0);
+    }
+
+    #[test]
+    fn blast_radius_direct() {
+        let modules = vec![make_module("a", "src/a.rs"), make_module("b", "src/b.rs")];
+        let deps = vec![Dependency {
+            from_module_id: "a".to_string(),
+            to_module_id: "b".to_string(),
+            line_number: 1,
+        }];
+        let hotspots = compute_hotspots(&modules, &deps);
+        let b = hotspots.iter().find(|h| h.path == "src/b.rs").unwrap();
+        assert_eq!(b.blast_radius, 1);
+    }
+
+    #[test]
+    fn blast_radius_transitive() {
+        let modules = vec![
+            make_module("a", "src/a.rs"),
+            make_module("b", "src/b.rs"),
+            make_module("c", "src/c.rs"),
+        ];
+        let deps = vec![
+            Dependency {
+                from_module_id: "a".to_string(),
+                to_module_id: "b".to_string(),
+                line_number: 1,
+            },
+            Dependency {
+                from_module_id: "b".to_string(),
+                to_module_id: "c".to_string(),
+                line_number: 1,
+            },
+        ];
+        let hotspots = compute_hotspots(&modules, &deps);
+        let c = hotspots.iter().find(|h| h.path == "src/c.rs").unwrap();
+        assert_eq!(c.blast_radius, 2);
+    }
+}
