+++ src/analyzer/tests.rs
@@ -407,221 +407,6 @@ fn detects_transitive_cycle() {
     );
 }
 
-#[test]
-fn independence_fully_internal() {
-    let modules = vec![
-        make_module("a1", "app/main.rs"),
-        make_module("a2", "app/util.rs"),
-    ];
-    let deps = vec![Dependency {
-        from_module_id: "a1".to_string(),
-        to_module_id: "a2".to_string(),
-        line_number: 1,
-    }];
-    let result = audit(&modules, &deps);
-    let app = result.independence.iter().find(|m| m.dir == "app");
-    assert!(app.is_some());
-    let app = app.unwrap();
-    assert_eq!(app.internal_deps, 1);
-    assert_eq!(app.external_deps, 0);
-    assert!((app.independence - 1.0).abs() < 0.01);
-}
-
-#[test]
-fn independence_mixed_deps() {
-    let modules = vec![
-        make_module("a1", "app/main.rs"),
-        make_module("a2", "app/util.rs"),
-        make_module("l1", "lib/core.rs"),
-    ];
-    let deps = vec![
-        Dependency {
-            from_module_id: "a1".to_string(),
-            to_module_id: "a2".to_string(),
-            line_number: 1,
-        },
-        Dependency {
-            from_module_id: "a1".to_string(),
-            to_module_id: "l1".to_string(),
-            line_number: 2,
-        },
-    ];
-    let result = audit(&modules, &deps);
-    let app = result.independence.iter().find(|m| m.dir == "app");
-    assert!(app.is_some());
-    let app = app.unwrap();
-    assert_eq!(app.internal_deps, 1);
-    assert_eq!(app.external_deps, 1);
-    assert!((app.independence - 0.5).abs() < 0.01);
-}
-
-#[test]
-fn independence_sorted_lowest_first() {
-    let modules = vec![
-        make_module("a1", "app/main.rs"),
-        make_module("a2", "app/util.rs"),
-        make_module("l1", "lib/core.rs"),
-        make_module("l2", "lib/helper.rs"),
-    ];
-    let deps = vec![
-        Dependency {
-            from_module_id: "a1".to_string(),
-            to_module_id: "a2".to_string(),
-            line_number: 1,
-        },
-        Dependency {
-            from_module_id: "a1".to_string(),
-            to_module_id: "l1".to_string(),
-            line_number: 2,
-        },
-        Dependency {
-            from_module_id: "l1".to_string(),
-            to_module_id: "l2".to_string(),
-            line_number: 1,
-        },
-    ];
-    let result = audit(&modules, &deps);
-    assert_eq!(result.independence.len(), 2);
-    assert_eq!(result.independence[0].dir, "app");
-    assert_eq!(result.independence[1].dir, "lib");
-}
-
-#[test]
-fn instability_pure_consumer() {
-    let modules = vec![
-        make_module("a1", "src/app/main.rs"),
-        make_module("l1", "src/lib/core.rs"),
-    ];
-    let deps = vec![Dependency {
-        from_module_id: "a1".to_string(),
-        to_module_id: "l1".to_string(),
-        line_number: 1,
-    }];
-    let result = audit(&modules, &deps);
-    let a1 = result
-        .hotspots
-        .iter()
-        .find(|h| h.path == "src/app/main.rs")
-        .unwrap();
-    assert_eq!(a1.fan_out, 1);
-    assert_eq!(a1.fan_in, 0);
-    assert!((a1.instability - 1.0).abs() < 0.01);
-}
-
-#[test]
-fn instability_pure_provider() {
-    let modules = vec![
-        make_module("a1", "src/app/main.rs"),
-        make_module("l1", "src/lib/core.rs"),
-    ];
-    let deps = vec![Dependency {
-        from_module_id: "a1".to_string(),
-        to_module_id: "l1".to_string(),
-        line_number: 1,
-    }];
-    let result = audit(&modules, &deps);
-    let l1 = result
-        .hotspots
-        .iter()
-        .find(|h| h.path == "src/lib/core.rs")
-        .unwrap();
-    assert_eq!(l1.fan_in, 1);
-    assert_eq!(l1.fan_out, 0);
-    assert!((l1.instability - 0.0).abs() < 0.01);
-}
-
-#[test]
-fn instability_balanced() {
-    let modules = vec![
-        make_module("a", "src/a.rs"),
-        make_module("b", "src/b.rs"),
-        make_module("c", "src/c.rs"),
-    ];
-    let deps = vec![
-        Dependency {
-            from_module_id: "a".to_string(),
-            to_module_id: "b".to_string(),
-            line_number: 1,
-        },
-        Dependency {
-            from_module_id: "b".to_string(),
-            to_module_id: "c".to_string(),
-            line_number: 1,
-        },
-    ];
-    let result = audit(&modules, &deps);
-    let b = result
-        .hotspots
-        .iter()
-        .find(|h| h.path == "src/b.rs")
-        .unwrap();
-    assert_eq!(b.fan_in, 1);
-    assert_eq!(b.fan_out, 1);
-    assert!((b.instability - 0.5).abs() < 0.01);
-}
-
-#[test]
-fn blast_radius_leaf_is_zero() {
-    let modules = vec![make_module("a", "src/a.rs"), make_module("b", "src/b.rs")];
-    let deps = vec![Dependency {
-        from_module_id: "a".to_string(),
-        to_module_id: "b".to_string(),
-        line_number: 1,
-    }];
-    let result = audit(&modules, &deps);
-    let a = result
-        .hotspots
-        .iter()
-        .find(|h| h.path == "src/a.rs")
-        .unwrap();
-    assert_eq!(a.blast_radius, 0);
-}
-
-#[test]
-fn blast_radius_direct() {
-    let modules = vec![make_module("a", "src/a.rs"), make_module("b", "src/b.rs")];
-    let deps = vec![Dependency {
-        from_module_id: "a".to_string(),
-        to_module_id: "b".to_string(),
-        line_number: 1,
-    }];
-    let result = audit(&modules, &deps);
-    let b = result
-        .hotspots
-        .iter()
-        .find(|h| h.path == "src/b.rs")
-        .unwrap();
-    assert_eq!(b.blast_radius, 1);
-}
-
-#[test]
-fn blast_radius_transitive() {
-    let modules = vec![
-        make_module("a", "src/a.rs"),
-        make_module("b", "src/b.rs"),
-        make_module("c", "src/c.rs"),
-    ];
-    let deps = vec![
-        Dependency {
-            from_module_id: "a".to_string(),
-            to_module_id: "b".to_string(),
-            line_number: 1,
-        },
-        Dependency {
-            from_module_id: "b".to_string(),
-            to_module_id: "c".to_string(),
-            line_number: 1,
-        },
-    ];
-    let result = audit(&modules, &deps);
-    let c = result
-        .hotspots
-        .iter()
-        .find(|h| h.path == "src/c.rs")
-        .unwrap();
-    assert_eq!(c.blast_radius, 2);
-}
-
 // ── audit_with_settings: the deterministic settings-aware seam ──
 
 #[test]
